% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMSD.R
\name{ELMSD}
\alias{ELMSD}
\title{Compressed representation of extreme learning machine with self-diffusion  network denoising applied on cancer subtyping.}
\usage{
ELMSD(x, nhid, k)
}
\arguments{
\item{x}{Gene expression matrix,Rows are samples and columns are features.}

\item{nhid}{Numbers of hidden nodes in extreme learning machine.}

\item{k}{Clustering number}
}
\value{
a list;
(1)clu:the clustering result;
(2)S:the final diffusion matrix;
}
\description{
Cancer subtypes identifition by dimensioanl reduction and denoising.
}
\examples{
data(Gene_expression)
test<-MOSD(Gene_expression,500,k)
}
